added ProductContext, CartContext files to manage the state of the products. Added basic cart functionality
