+++ package.json
@@ -32,6 +32,7 @@
     "react-bootstrap-icons": "^1.2.3",
     "react-dom": "^17.0.1",
     "react-router-dom": "^5.2.0",
+    "react-router-hash-link": "^2.3.1",
     "react-scripts": "4.0.1",
     "web-vitals": "^0.2.4"
   },
