+++ src/App.js
@@ -1,21 +1,27 @@
-import { Route, Link, BrowserRouter as Router, Switch } from "react-router-dom";
+import {ProductProvider} from './ProductContext'
+import {CartProvider} from './CartContext'
 import Navigation from "./components/NavigationBar/Navigation";
-import HeroSection from "./components/HomeSection/HeroSection";
-import InfoSection from "./components/FaqSection/InfoSection";
-import ContactSection from "./components/ContactFormSection/ContactSection";
-import LoginSection from "./components/LoginSection/LoginSection";
-import ProductSection from "./components/ProductSection/ProductSection";
-import AboutUs from "./components/AboutUsSection/AboutUsSection";
-import FaqSection from "./components/FaqSection/FaqSection";
-import Impressum  from "./components/ImpressumExtraPage/Impressum";
+// import HeroSection from "./components/HomeSection/HeroSection";
+// import InfoSection from "./components/FaqSection/InfoSection";
+// import ContactSection from "./components/ContactFormSection/ContactSection";
+// import LoginSection from "./components/LoginSection/LoginSection";
+// import ProductSection from "./components/ProductSection/ProductSection";
+// import AboutUs from "./components/AboutUsSection/AboutUsSection";
+// import FaqSection from "./components/FaqSection/FaqSection";
+// import Impressum  from "./components/ImpressumExtraPage/Impressum";
 
 import "./App.scss";
 
 function App() {
   return (
+    <ProductProvider>
+    <CartProvider>
     <div className="App">
       <Navigation />
+      
     </div>
+    </CartProvider>
+    </ProductProvider>
   );
 }
