+++ src/components/Home.js
@@ -1,17 +1,13 @@
+import {useState} from 'react'
 import tree from "../tree.json";
+import TreeCard from "./ProductSection/TreeCard";
 import ProductSection from "../components/ProductSection/ProductSection";
-import { withRouter } from "react-router-dom";
 
 function Home(props) {
-  const { location } = props;
-  function getState(state) {
-    return state;
-  }
 
   return (
     <div className="App">
-      <ProductSection products={tree} state={getState} />
-      {location.pathname}
+      <ProductSection products={tree} />
     </div>
   );
 }
